% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{计算矩阵之间的距离}
\usage{
distance(M, ...)
}
\arguments{
\item{M}{初始化生成的相对矩阵}
}
\value{
\code{M} 矩阵中点的相对距离，返回一个 n x n的距离矩阵
}
\description{
计算矩阵之间的距离
}
\examples{
initpos(14, 5, 5)
M  <- initpos(n = 14, ncol = 5, nrow = 5)
distance(M)
}
\author{
muzhiyang
}
