% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logie.optimal.R
\name{logie.optimal}
\alias{logie.optimal}
\title{优化位置矩阵}
\usage{
logie.optimal(M, S, method = "HA", HA.N = 100)
}
\arguments{
\item{M}{位置矩阵}

\item{S}{综合因素矩阵}

\item{N}{计算的循环次数}
}
\value{
返回一个\code{list}的数据结构
\code{list(result_cost, result_M )}
包含得到的较优解记录和位置矩阵的记录

  costfunction to evaule the method.
}
\description{
优化位置矩阵
}
